refactored base button and default text button
