+++ lib/app/auth/presentation/pages/login/widgets/login_section/about_button.dart
@@ -1,6 +1,7 @@
 import 'package:crypto_journal_mobile/shared/constants/constants.dart';
 import 'package:crypto_journal_mobile/shared/functions/launch_url.dart';
 import 'package:crypto_journal_mobile/shared/theme/constants.dart';
+import 'package:crypto_journal_mobile/shared/widgets/buttons/base_button.dart';
 import 'package:crypto_journal_mobile/shared/widgets/buttons/default_text_button.dart';
 import "package:flutter/material.dart";
