+++ lib/app/auth/presentation/pages/login/widgets/login_section/sign_in_with_google_button.dart
@@ -1,6 +1,7 @@
 import 'package:crypto_journal_mobile/app/auth/presentation/pages/login/functions/sign_in.dart';
 import 'package:crypto_journal_mobile/app/auth/service/dtos/sign_in_dto.dart';
 import 'package:crypto_journal_mobile/shared/theme/colors.dart';
+import 'package:crypto_journal_mobile/shared/widgets/buttons/base_button.dart';
 import 'package:crypto_journal_mobile/shared/widgets/buttons/default_text_button.dart';
 import "package:flutter/material.dart";
