+++ lib/app/auth/presentation/pages/settings/widgets/log_out_button.dart
@@ -1,5 +1,6 @@
 import 'package:crypto_journal_mobile/app/auth/presentation/pages/settings/functions/sign_out.dart';
 import 'package:crypto_journal_mobile/shared/theme/colors.dart';
+import 'package:crypto_journal_mobile/shared/widgets/buttons/base_button.dart';
 import 'package:crypto_journal_mobile/shared/widgets/buttons/default_text_button.dart';
 import "package:flutter/material.dart";
