+++ lib/app/user/presentation/pages/home/home_page.dart
@@ -36,7 +36,6 @@ class HomePage extends StatelessWidget {
                   bottom: defaultPagePadding,
                 ),
                 child: BaseButton(
-                  backGroundColor: backGroundColorSecondary,
                   child: TabBar(
                     indicatorColor: Colors.transparent,
                     unselectedLabelColor: textColorSecondary,
