+++ lib/shared/widgets/buttons/default_text_button.dart
@@ -1,15 +1,7 @@
 import 'package:crypto_journal_mobile/shared/theme/constants.dart';
-import 'package:crypto_journal_mobile/shared/theme/container_decoration.dart';
 import 'package:crypto_journal_mobile/shared/theme/text_styles.dart';
-import 'package:crypto_journal_mobile/shared/widgets/containers/base_layout_container.dart';
+import 'package:crypto_journal_mobile/shared/widgets/buttons/base_button.dart';
 import "package:flutter/material.dart";
-
-enum ButtonWidth {
-  Full,
-  Medium,
-  Half,
-  Small,
-}
 
 class DefaultTextButton extends StatelessWidget {
   final void Function()? onTap;
@@ -39,53 +31,22 @@ class DefaultTextButton extends StatelessWidget {
     );
   }
 
-  BoxDecoration? _boxDecorationBuilder() {
-    final decoration = this.color == null
-        ? defaultContainerDecoration
-        : defaultContainerDecoration.copyWith(
-            color: this.color,
-          );
-
-    return this.hasDecoration ? decoration : null;
-  }
-
-  double _buttonWidthBuilder(double width) {
-    switch (this.width) {
-      case ButtonWidth.Full:
-        return width;
-      case ButtonWidth.Medium:
-        return width * 3 / 4;
-      case ButtonWidth.Half:
-        return width / 2;
-      case ButtonWidth.Small:
-        return width / 4;
-      default:
-        return width / 2;
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
-    return BaseLayoutContainer(
-      heigthProp: buttonHeight,
-      builder: (BuildContext context, Size size) {
+    return BaseButton(
+      color: this.color,
+      hasDecoration: this.hasDecoration,
+      key: this.key,
+      onTap: this.onTap,
+      width: this.width,
+      builder: (context, size) {
         final height = size.height;
 
-        return Align(
-          child: GestureDetector(
-            onTap: this.onTap ?? () {},
-            child: Container(
-              alignment: Alignment.center,
-              width: this._buttonWidthBuilder(size.width),
-              child: Center(
-                child: Text(
-                  this.text,
-                  style: this._textStyleBuilder(
-                    height,
-                  ),
-                ),
-              ),
-              decoration: this._boxDecorationBuilder(),
+        return Center(
+          child: Text(
+            this.text,
+            style: this._textStyleBuilder(
+              height,
             ),
           ),
         );
